Recommendation System/Movie_Recommendation/requirements.txt
streamlit
pandas
requests
langchain
chromadb
langchain_huggingface
scikit-learn
sentence-transformers
joblib
gunicorn
langchain-chroma
pinecone-client
